+++ AlfrescoSDK/AlfrescoSDK/Utils/AlfrescoFavoritesCache.h
@@ -34,9 +34,9 @@ typedef enum
 @property (nonatomic, assign, readonly) BOOL hasMoreFavoriteDocuments;
 @property (nonatomic, assign, readonly) BOOL hasMoreFavoriteFolders;
 @property (nonatomic, assign, readonly) BOOL hasMoreFavoriteNodes;
-@property (nonatomic, assign, readonly) NSInteger totalDocuments;
-@property (nonatomic, assign, readonly) NSInteger totalFolders;
-@property (nonatomic, assign, readonly) NSInteger totalNodes;
+@property (nonatomic, assign, readonly) int totalDocuments;
+@property (nonatomic, assign, readonly) int totalFolders;
+@property (nonatomic, assign, readonly) int totalNodes;
 /**
  initialiser
  */
